moving proxy worker around
